Add policies, add gate authorize to controllers, add gate define in AppServiceProvider

+++ api/app/Http/Controllers/OrgController.php
@@ -12,8 +12,6 @@ use App\Traits\IncludeRelationships;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Gate;
-use Illuminate\Support\Facades\Log;
 
 class OrgController extends Controller
 {
@@ -21,8 +19,7 @@ class OrgController extends Controller
 
     public function index(OrgFilter $filter, Request $request): OrgCollection
     {
-        Gate::authorize('viewAny', Org::class);
-        
+        $this->authorize('viewAny', Org::class);        
         $org = Org::filter($filter);
 
         return new OrgCollection(
@@ -32,6 +29,7 @@ class OrgController extends Controller
 
     public function store(StoreOrgRequest $request): OrgResource
     {
+        $this->authorize('create', Org::class);
         $validated = $request->validated();
         $org = Org::create($validated);
 
@@ -40,14 +38,18 @@ class OrgController extends Controller
 
     public function show(string $id): OrgResource
     {
-        $org = Org::query();
-        $this->applyIncludes($org, request());
+        
+        $orgQuery = Org::query();
+        $this->applyIncludes($orgQuery, request());
+        $org = $orgQuery->findOrFail($id);
+        $this->authorize('view', $org);
 
-        return new OrgResource($org->findOrFail($id));
+        return new OrgResource($org);
     }
 
     public function update(UpdateOrgRequest $request, Org $org): OrgResource
     {
+        $this->authorize('update', $org);
         $validated = $request->validated();
         $org->update($validated);
 
@@ -56,6 +58,7 @@ class OrgController extends Controller
 
     public function destroy(Org $org): Response
     {
+        $this->authorize('delete', $org);
         $org->deleted_by = Auth::id();
         $org->save();
         $org->delete();
